requirements.txt
django
dnspython